+++ scanpipe/tests/test_commands.py
@@ -49,6 +49,7 @@ from scanpipe.models import WebhookSubscription
 from scanpipe.pipes import flag
 from scanpipe.pipes import purldb
 from scanpipe.tests import filter_warnings
+from scanpipe.tests import make_mock_response
 from scanpipe.tests import make_package
 from scanpipe.tests import make_project
 from scanpipe.tests import make_resource_file
@@ -319,7 +320,7 @@ class ScanPipeManagementCommandTest(TestCase):
     def test_scanpipe_management_command_add_input_file(self):
         out = StringIO()
 
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         parent_path = Path(__file__).parent
         options = [
             "--input-file",
@@ -346,7 +347,7 @@ class ScanPipeManagementCommandTest(TestCase):
             call_command("add-input", *options, stdout=out)
 
     def test_scanpipe_management_command_add_input_url(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         options = [
             "--input-url",
             "https://example.com/archive.zip",
@@ -368,7 +369,7 @@ class ScanPipeManagementCommandTest(TestCase):
     def test_scanpipe_management_command_add_input_copy_codebase(self):
         out = StringIO()
 
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
 
         options = ["--copy-codebase", "non-existing", "--project", project.name]
         expected = "non-existing not found"
@@ -395,7 +396,7 @@ class ScanPipeManagementCommandTest(TestCase):
     def test_scanpipe_management_command_add_pipeline(self):
         out = StringIO()
 
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
 
         pipelines = [
             self.pipeline_name,
@@ -432,7 +433,7 @@ class ScanPipeManagementCommandTest(TestCase):
     def test_scanpipe_management_command_add_webhook(self):
         out = StringIO()
 
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
 
         options = ["https://example.com/webhook"]
         expected = "the following arguments are required: --project"
@@ -487,7 +488,7 @@ class ScanPipeManagementCommandTest(TestCase):
             "analyze_root_filesystem_or_vm_image",
         ]
 
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         for pipeline_name in pipeline_names:
             project.add_pipeline(pipeline_name)
 
@@ -514,7 +515,7 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertEqual(expected, out.getvalue())
 
     def test_scanpipe_management_command_execute(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         options = ["--project", project.name]
 
         out = StringIO()
@@ -558,7 +559,7 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertEqual("", run3.task_output)
 
     def test_scanpipe_management_command_execute_project_function(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
 
         expected = "No pipelines to run on project my_project"
         with self.assertRaisesMessage(CommandError, expected):
@@ -585,7 +586,7 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertIsNone(returned_value)
 
     def test_scanpipe_management_command_status(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         run = project.add_pipeline(self.pipeline_name)
 
         options = ["--project", project.name, "--no-color"]
@@ -658,9 +659,9 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertIn("(addon)", output)
 
     def test_scanpipe_management_command_list_project(self):
-        project1 = Project.objects.create(name="project1")
-        project2 = Project.objects.create(name="project2")
-        project3 = Project.objects.create(name="archived", is_archived=True)
+        project1 = make_project(name="project1")
+        project2 = make_project(name="project2")
+        project3 = make_project(name="archived", is_archived=True)
 
         options = []
         out = StringIO()
@@ -687,7 +688,7 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertIn(project3.name, output)
 
     def test_scanpipe_management_command_output(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         make_package(project, package_url="pkg:generic/name@1.0")
 
         out = StringIO()
@@ -755,7 +756,7 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertIn('"specVersion": "1.5",', out_value)
 
     def test_scanpipe_management_command_delete_project(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         work_path = project.work_path
         self.assertTrue(work_path.exists())
 
@@ -771,7 +772,7 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertFalse(work_path.exists())
 
     def test_scanpipe_management_command_archive_project(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         (project.input_path / "input_file").touch()
         (project.codebase_path / "codebase_file").touch()
         self.assertEqual(1, len(Project.get_root_content(project.input_path)))
@@ -798,7 +799,7 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertEqual(0, len(Project.get_root_content(project.codebase_path)))
 
     def test_scanpipe_management_command_reset_project(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
         project.add_pipeline("analyze_docker_image")
         CodebaseResource.objects.create(project=project, path="filename.ext")
         DiscoveredPackage.objects.create(project=project)
@@ -834,8 +835,8 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertEqual(0, len(Project.get_root_content(project.codebase_path)))
 
     def test_scanpipe_management_command_flush_projects(self):
-        project1 = Project.objects.create(name="project1")
-        project2 = Project.objects.create(name="project2")
+        project1 = make_project("project1")
+        project2 = make_project("project2")
         ten_days_ago = timezone.now() - datetime.timedelta(days=10)
         project2.update(created_date=ten_days_ago)
 
@@ -847,12 +848,56 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertEqual(expected, out_value)
         self.assertEqual(project1, Project.objects.get())
 
-        Project.objects.create(name="project2")
+        make_project("project2")
+        out = StringIO()
+        options = ["--no-color", "--no-input", "--dry-run"]
+        call_command("flush-projects", *options, stdout=out)
+        out_value = out.getvalue().strip()
+        expected = "2 projects would be deleted:\n- project2\n- project1"
+        self.assertEqual(expected, out_value)
+
         out = StringIO()
         options = ["--no-color", "--no-input"]
         call_command("flush-projects", *options, stdout=out)
         out_value = out.getvalue().strip()
         expected = "2 projects and their related data have been removed."
+        self.assertEqual(expected, out_value)
+
+    def test_scanpipe_management_command_flush_projects_filters(self):
+        label1 = "label1"
+        label2 = "label2"
+        make_project("project1", labels=[label1])
+        make_project("project2", labels=[label1, label2])
+        make_project("project3", pipelines=["scan_single_package"])
+
+        base_options = ["--no-color", "--no-input", "--dry-run"]
+
+        out = StringIO()
+        options = base_options + ["--label", label1]
+        call_command("flush-projects", *options, stdout=out)
+        out_value = out.getvalue().strip()
+        expected = "2 projects would be deleted:\n- project2\n- project1"
+        self.assertEqual(expected, out_value)
+
+        out = StringIO()
+        options = base_options + ["--label", label2]
+        call_command("flush-projects", *options, stdout=out)
+        out_value = out.getvalue().strip()
+        expected = "1 projects would be deleted:\n- project2"
+        self.assertEqual(expected, out_value)
+
+        out = StringIO()
+        options = base_options + ["--label", label1, "--label", label2]
+        call_command("flush-projects", *options, stdout=out)
+        out_value = out.getvalue().strip()
+        expected = "2 projects would be deleted:\n- project2\n- project1"
+        self.assertEqual(expected, out_value)
+
+        out = StringIO()
+        options = base_options + ["--pipeline", "scan_single_package"]
+        call_command("flush-projects", *options, stdout=out)
+        out_value = out.getvalue().strip()
+        expected = "1 projects would be deleted:\n- project3"
         self.assertEqual(expected, out_value)
 
     def test_scanpipe_management_command_create_user(self):
@@ -965,12 +1010,7 @@ class ScanPipeManagementCommandTest(TestCase):
         mock_get_latest_output.return_value = (
             self.data / "scancode" / "is-npm-1.0.0_summary.json"
         )
-        mock_download_get.return_value = mock.Mock(
-            content=b"\x00",
-            headers={},
-            status_code=200,
-            url=download_url,
-        )
+        mock_download_get.return_value = make_mock_response(url=download_url)
 
         self.assertFalse(WebhookSubscription.objects.exists())
 
@@ -1018,12 +1058,7 @@ class ScanPipeManagementCommandTest(TestCase):
             "status": f"updated scannable_uri {scannable_uri_uuid} "
             "scan_status to 'failed'"
         }
-        mock_download_get.return_value = mock.Mock(
-            content=b"\x00",
-            headers={},
-            status_code=200,
-            url=download_url,
-        )
+        mock_download_get.return_value = make_mock_response(url=download_url)
 
         options = [
             "--max-loops",
@@ -1077,18 +1112,8 @@ class ScanPipeManagementCommandTest(TestCase):
         ]
 
         mock_download_get.side_effect = [
-            mock.Mock(
-                content=b"\x00",
-                headers={},
-                status_code=200,
-                url=download_url1,
-            ),
-            mock.Mock(
-                content=b"\x00",
-                headers={},
-                status_code=200,
-                url=download_url2,
-            ),
+            make_mock_response(url=download_url1),
+            make_mock_response(url=download_url2),
         ]
 
         mock_request_post.side_effect = [
@@ -1144,7 +1169,7 @@ class ScanPipeManagementCommandTest(TestCase):
         )
 
     def test_scanpipe_management_command_check_compliance(self):
-        project = Project.objects.create(name="my_project")
+        project = make_project(name="my_project")
 
         out = StringIO()
         options = ["--project", project.name]
@@ -1190,9 +1215,8 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertEqual(expected, out_value)
 
     def test_scanpipe_management_command_report(self):
-        project1 = make_project("project1")
         label1 = "label1"
-        project1.labels.add(label1)
+        project1 = make_project("project1", labels=[label1])
         make_resource_file(project1, path="file.ext", status=flag.REQUIRES_REVIEW)
         make_project("project2")